文件存储改为七牛云存储

+++ app/Carousel.php
@@ -4,6 +4,8 @@ namespace App;
 
 
 use Illuminate\Database\Eloquent\Model;
+use Qiniu\Auth;
+use Qiniu\Storage\UploadManager;
 
 class Carousel extends Model
 {
@@ -11,11 +13,21 @@ class Carousel extends Model
         'order', 'url',
     ];
     protected $table = 'carousels';
+    private $accessKey = 'yi-qu1G_W7fSnAcH2GiLvg4BIbB0Bu2swKBXW_P8';
+    private $secretKey = 'Nu1ntfUCCBkPEVQYMZfi2Pvsb0VqBefecvjlNQu2';
     public function add($carousel,$order){
-        $path = $carousel->storeAS(
-            'carousels',uniqid().'.jpg'
-        );
-        $url = 'http://www.thmaoqiu.cn/saiyou/storage/app/'.$path;
+        $auth = new Auth($this->accessKey, $this->secretKey);
+        $bucket = 'maoqiu';
+        // 生成上传Token
+        $token = $auth->uploadToken($bucket);
+        // 构建 UploadManager 对象
+        $uploadMgr = new UploadManager();
+        $key = uniqid().'.jpg';  //自动生成的文件名
+        if ($uploadMgr->putFile($token,$key,$carousel)){
+            $url = 'http://otq91javs.bkt.clouddn.com/'.$key;
+        }else{
+            return response()->json(['code'=>5,'msg'=>'七牛云连接失败']);
+        }
         $carousels = new Carousel();
         $carousels->url = $url;
         $carousels->order = $order;
@@ -29,11 +41,23 @@ class Carousel extends Model
     public function del($order){
         $carousels = new Carousel();
         if ($carousels->where('order',$order)->first()){
-            if ($carousels->where('order',$order)->delete()){
-                return response()->json(['code'=>0,'msg'=>'删除轮播图成功']);
+
+            $picNames = explode('/',$carousels->url);
+            $picName = $picNames[sizeof($picNames)-1];
+            $auth = new Auth($this->accessKey, $this->secretKey);
+            $bucket = 'maoqiu';
+            $config = new \Qiniu\Config();
+            $bucketManager = new \Qiniu\Storage\BucketManager($auth, $config);
+            if ($bucketManager->delete($bucket, $picName)){
+                if ($carousels->where('order',$order)->delete()){
+                    return response()->json(['code'=>0,'msg'=>'删除轮播图成功']);
+                }else{
+                    return response()->json(['code'=>1,'msg'=>'删除轮播图失败']);
+                }
             }else{
-                return response()->json(['code'=>1,'msg'=>'删除轮播图失败']);
+                return response()->json(['code'=>5,'msg'=>'七牛云连接失败']);
             }
+
         }else{
             return response()->json(['code'=>4,'msg'=>'轮播图未找到']);
         }
